third_party/blink/web_tests/svg/dom/systemLanguage-comma-separated.html

<!DOCTYPE HTML>
<title>The "systemLanguage" is a COMMA-separated list of langtags</title>
<link rel="help" href="https://www.w3.org/TR/SVG2/struct.html#ConditionalProcessingSystemLanguageAttribute">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg>
	<text systemLanguage="en-US, zh-Hans,zh-Hant" id="switchtest"></text>
</svg>
<script>
test(() => {
	var systemLanguage_langtagList = document.getElementById("switchtest").systemLanguage;
	assert_equals(systemLanguage_langtagList.numberOfItems, 3, 'The number of langtag should be 3');
	assert_equals(systemLanguage_langtagList.getItem(0), 'en-US');
	assert_equals(systemLanguage_langtagList.getItem(1), 'zh-Hans');
	assert_equals(systemLanguage_langtagList.getItem(2), 'zh-Hant');
});

</script>
